Refactored the Search app to contain all rest code in the "rest_framework" folder
