+++ api/app/signals/apps/search/rest_framework/pagination.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
 from datapunt_api.pagination import HALPagination
 from django.core.paginator import Page, Paginator
 from rest_framework.exceptions import NotFound
